LayoutTests/http/tests/security/XFrameOptions/x-frame-options-deny-meta-tag-in-body.html

<script>
    if (window.layoutTestController) {
        layoutTestController.dumpAsText();
        layoutTestController.dumpChildFramesAsText();
        layoutTestController.dumpResourceLoadCallbacks();
        layoutTestController.waitUntilDone();
    }

    function checkIfDone()
    {
        if (document.getElementsByTagName("iframe")[0].contentWindow.location == "about:blank")
            layoutTestController.notifyDone();
    }
</script>

<p>There should be no content in the iframe below</p>
<iframe style="width:500px; height:500px" src="http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny-meta-tag-subframe-in-body.html" onload="checkIfDone()"></iframe>
